Move Continuable jobs into the test case
