--- activejob/test/jobs/continuable_duplicate_step_job.rb
@@ -1,12 +1,0 @@
-# frozen_string_literal: true
-
-class ContinuableDuplicateStepJob < ActiveJob::Base
-  include ActiveJob::Continuable
-
-  def perform
-    step :duplicate do |step|
-    end
-    step :duplicate do |step|
-    end
-  end
-end
