--- activejob/test/jobs/continuable_string_step_name_job.rb
@@ -1,10 +1,0 @@
-# frozen_string_literal: true
-
-class ContinuableStringStepNameJob < ActiveJob::Base
-  include ActiveJob::Continuable
-
-  def perform
-    step "string_step_name" do |step|
-    end
-  end
-end
